cms-admin: Restructure action log module for multiple views
Reshape src/actionLog into a clearer two-folder layout to make room
for upcoming action log views that share the same dumb building
blocks:

- src/actionLog/components/  -  shared dumb components used by every
                                view (ActionLogCompare,
                                ActionLogShowVersion, ActionLogHeader,
                                DiffHeader, DiffViewer, UserCell)
- src/actionLog/actionLog/   -  per-record (single entity) components
                                (ActionLogButton, ActionLogDialog,
                                ActionLogVersionGrid)

Per-record renames in line with what they actually show:
- ActionLogGrid              -> ActionLogVersionGrid
- actionLogGridFragment      -> actionLogVersionGridFragment
- i18n keys actionLog.actionLogGrid.* -> actionLog.actionLogVersionGrid.*

Other cleanups:
- Co-locate the snapshot key filter helper next to its only owner
  (actionLog.utils.ts -> components/actionLogCompare/ActionLogCompare.utils.ts)
- DiffHeader gets its own folder, matching every other components/ sibling
- Rename components/header -> components/actionLogHeader so the folder
  name matches the component file
- Translate the remaining German labels in the per-record grid +
  toolbar to English
- Align Storybook titles to path-case across the module

No behavior change.
