--- packages/admin/cms-admin/src/actionLog/actionLogGrid/ActionLogGrid.gql.ts
@@ -1,14 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const actionLogGridFragment = gql`
-    fragment ActionLogGrid on ActionLog {
-        id
-        user {
-            id
-            name
-        }
-        entityName
-        version
-        createdAt
-    }
-`;
