--- packages/admin/cms-admin/src/actionLog/components/__stories__/DiffHeader.stories.tsx
@@ -1,27 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react-vite";
-
-import { DiffHeader } from "../DiffHeader";
-
-type Story = StoryObj<typeof DiffHeader>;
-const meta: Meta<typeof DiffHeader> = {
-    component: DiffHeader,
-    title: "actionLog/components/DiffHeader",
-};
-export default meta;
-
-export const StandardDiffHeader: Story = {
-    args: {
-        createdAt: "2023-10-01T12:00:00Z",
-        userId: "abc-123",
-        userName: "Max Mustermann",
-        version: 1,
-    },
-};
-
-export const UnknownUser: Story = {
-    args: {
-        createdAt: "2023-10-01T12:00:00Z",
-        userId: "abc-123",
-        version: 1,
-    },
-};
